Generate record contracts for some JSON schemas
This is a first attempt at generating Nickel record contracts instead of
predicates for some JSON schemas. It only works for schemas that specify
an `instance_type` of `"object"` and contain only the JSON schema
keywords `properties` and `required`. For `additionalProperties` only a
boolean schema `true` or `false` is supported.