Generate record contracts for some JSON schemas (#3)
* Generate record contracts for some JSON schemas

This is a first attempt at generating Nickel record contracts instead of
predicates for some JSON schemas. It only works for schemas that specify
an `instance_type` of `"object"` and contain only the JSON schema
keywords `properties` and `required`. For `additionalProperties` only a
boolean schema `true` or `false` is supported.

* Split `src/lib.rs` for predicate and contract handling

* Start threading an environment of definitions

* Handle references to subcontracts without falling back to predicates

* Move test schema into subdirectory

* Rename `open_record`

Co-authored-by: Yann Hamdaoui <yann.hamdaoui@tweag.io>

* Get rid of `EnvironmentF` in favor of direct newtypes.

* Use `Default::default()` for constructing Nickel `Field`s

* Document the code generation approach

---------

Co-authored-by: Yann Hamdaoui <yann.hamdaoui@tweag.io>