Remove the separate reference access environment (#7)
* Remove the separate reference access environment

The `References` type was originally introduced to track if particular
top-level definitions in a JSON schema could be turned into a record
contract or not. Accordingly there would be a `RichTerm` representing
a reference to the record contract form and one for the predicate
form. However, it turns out that we need to have access to both a
predicate and a contract form for all definitions anyway. This makes the
`References` environment superfluous since it's much easier to generate
the appropiate `RichTerm`s on the fly.

* Update documentation since we no longer have a field access environment