Relax conditions on `extensions` for contract generation (#14)
The first prototype checked if the `extensions` map in a `SchemaObject`
is empty before going ahead with contract generation. However, I have
seen JSON schemas use a `$comment` field sometimes. The `schemars` crate
does not recognize this field and treats it as an `extension`. This
PR introduces a concept of "ignored fields" in the extensions map and
proceeds with contract generation if only ignored fields are present.