Fix interpolation and quoting in predicate error messages
When the Nickel lexer was changed to parse `"%"` as a string ending
delimiter even when a `{` follows, I naively changed some string
delimiters to `m%%`. However, @thufschmitt correctly observed that this
way no interpolation will take place. This PR is a copout: I just use
backticks instead of double quotes and reintroduce the single-percent
delimiters.