Don't panic on external refs
Using external schemas as references is not supported by the current
version of json-schema-to-nickel. The previous implemented just
panicked, which is unfortunate as it's an all-or-nothing approach: if
there's a ref somewhere in a schema, you just can't do anything with it.

This commit adopts a middle ground by warning about unfollowed external
refs, and substitute them for `Dyn` contracts.