Extend reference handling to support properties
Previously, json-schema-to-nickel was only able to handle local
references to top-level definitions, that is references of the form
"#/definitions/xxx`. This commit is a first step toward handling
references to local properties as well, of the form
`#/properties/foo/properties/bar/properties/baz`.

Doing so, we also rework the way top-level definitions are handled as
well, for better uniformity and to reduce the size of generated
contracts (we use to include all definitions by default, while this
change pushes toward recording which ones are actually referenced, and
keep only those).