Re-enable some ref tests (#73)
Enable part of the test suite on references. For now, we've simply
re-enable the tests that are passing with the current state of the
repository, which at least mean those won't regress in future update. It
would be better to see, for each failing test, if it's failing because
it's unsupported or because there's a bug in the current implementation
- this is left for a future commit.