Add --library-path argument
By default, the internal js2n predicate library is inlined in each
generated contract to ensure that this is a standalone Nickel file that
can be imported and moved around freely. However, this eats quite a lot
of space, especially when a lot of contracts are generated and
distributed together.

This commits add a CLI argument, `--library-path`, to provide a path from
which the generated contract should import the library, instead of
inlining it, so that the library can be shared by many contracts.