Support percent-encoded URIs (#78)
This commit makes use of a proper URI parser, in particular to correctly
support JSON schema references with percent-encoded characters. This
might also prove useful if we ever want to resolve distant URIs at some
point.

The `definitions` module is also renamed to `references` to better
reflect the fact that it now manages general reference resolution, and
not just definitions. A bunch of warning messages have been marginally
improved.