Add a source data to Nickel error type
This commit threads additional data through predicates, namely the error
source in the configuration - we can't represent it as a Nickel
position, but similar to a JSON pointer, it's represented as a path of
nested field accesses and array indices. This is put to use by including
this information into error messages, leading to much better diagnostic
when facing validation errors that happen deep inside an object.