Improve contract errors (#112)
* WIP

* Add a source data to Nickel error type

This commit threads additional data through predicates, namely the error
source in the configuration - we can't represent it as a Nickel
position, but similar to a JSON pointer, it's represented as a path of
nested field accesses and array indices. This is put to use by including
this information into error messages, leading to much better diagnostic
when facing validation errors that happen deep inside an object.

* Update to latest release of Nickel - WIP

* Fix Rust compilation errors

* fix arrayOf plus formatting

* Use mk_error helper

* Fix infinite recursion in mk_error

* Fix order of adding element to error source, improve string representation

* Fix missing interpolation